backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
aiogram==2.25.1
python-dotenv==1.0.0
requests==2.31.0
cryptography==41.0.7
PyJWT==2.8.0
bcrypt==4.0.1
psycopg2-binary==2.9.7
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0 